Move FusedSDPA QKV slice to ModuleFusedSDPA
