+++ vllm_hpu_extension/utils.py
@@ -10,6 +10,8 @@ from functools import lru_cache, wraps
 
 import habana_frameworks.torch as htorch
 import torch
+from habana_frameworks.torch.hpex.kernels.FusedSDPA import (
+    gqa_input_reshape_fwd, gqa_output_reshape, is_gqa)
 
 from vllm_hpu_extension.runtime import get_config
 
@@ -147,6 +149,31 @@ class ModuleFusedSDPA(torch.nn.Module):
         assert fusedSDPA is not None, f'fusedSDPA kernel is None'
         self._hpu_kernel_fsdpa = fusedSDPA
 
+        impl_mapping = {
+            'split_kv': SplitPrefixCausalSDPA,
+            'slice_causal': CausalSliceSDPA,
+            'slice_qkv': QKVSliceSDPA,
+        }
+
+        sdpa_slice_impl_option = get_config().VLLM_HPU_FSDPA_SLICE_IMPL
+        sdpa_slice_impl_option = sdpa_slice_impl_option if sdpa_slice_impl_option is not None else 'slice_qkv'
+        assert sdpa_slice_impl_option in impl_mapping, f'Unsupported sdpa slice impl: {sdpa_slice_impl_option}'
+
+        self.sdpa_slice_impl = impl_mapping[sdpa_slice_impl_option]
+        self.slice_causal = get_config().VLLM_HPU_FSDPA_SLICE_CAUSAL
+
+        qkv_slice_thld = get_config().VLLM_HPU_FSDPA_SLICE_SEQ_LEN_THLD
+        self.qkv_slice_thld = qkv_slice_thld if qkv_slice_thld is not None else 0
+
+        q_chunk_size = get_config().VLLM_HPU_FSDPA_SLICE_CHUNK_SIZE
+        self.q_chunk_size = q_chunk_size if q_chunk_size is not None else 4096
+
+        kv_chunk_size = get_config().VLLM_HPU_FSDPA_SLICE_CHUNK_SIZE
+        self.kv_chunk_size = kv_chunk_size if kv_chunk_size is not None else 4096
+
+        causal_chunk_size = get_config().VLLM_HPU_FSDPA_SLICE_CHUNK_SIZE
+        self.causal_chunk_size = causal_chunk_size if causal_chunk_size is not None else 4096
+
     def forward(
         self,
         query,
@@ -163,8 +190,79 @@ class ModuleFusedSDPA(torch.nn.Module):
         window_size=None,
     ):
 
-        if window_size:
-            return self._hpu_kernel_fsdpa.apply(
+        bs = query.size(0)
+        query_len = query.size(-2)
+        kv_len = key.size(-2)
+
+        # prefill with prefix caching
+        if (bs == 1 and self.qkv_slice_thld > 0 and query_len != 1
+                and query_len != kv_len and kv_len >= self.qkv_slice_thld):
+            prefix_len = kv_len - query_len
+
+            key_prefix = key[..., :prefix_len, :]
+            value_prefix = value[..., :prefix_len, :]
+
+            key_causal = key[..., prefix_len:, :]
+            value_causal = value[..., prefix_len:, :]
+
+            assert attn_mask is not None, "attn_mask should not be None for prefix caching prefill"
+            causal_attn_mask = attn_mask[..., prefix_len:]
+
+            attn_weights = self.sdpa_slice_impl.apply(
+                query,
+                key_prefix,
+                value_prefix,
+                key_causal,
+                value_causal,
+                causal_attn_mask,
+                self.q_chunk_size,
+                self.kv_chunk_size,
+                self.causal_chunk_size,
+                scale,
+                softmax_mode,
+                None,  #valid_seq_len
+                padding_side)
+
+            return attn_weights
+
+        elif (is_causal and window_size is None and self.slice_causal
+              and self.qkv_slice_thld > 0
+              and query_len >= self.qkv_slice_thld):
+            attn_results = QKVSliceCausalSDPA.apply(
+                query,
+                key,
+                value,
+                self.causal_chunk_size,
+                attn_mask,
+                scale,
+                softmax_mode,
+                None,  #valid_seq_len
+                padding_side,
+                True  # output_original_dtype
+            )
+            attn_weights = attn_results[0]
+
+            return attn_weights
+
+        else:
+            if window_size:
+                return self._hpu_kernel_fsdpa.apply(
+                    query,
+                    key,
+                    value,
+                    attn_mask,
+                    dropout_p,
+                    is_causal,
+                    scale,
+                    softmax_mode,
+                    recompute_mode,
+                    valid_sequence_lengths,
+                    padding_side,
+                    False,
+                    False,
+                    window_size)
+            else:
+                return self._hpu_kernel_fsdpa.apply(
                 query,
                 key,
                 value,
@@ -176,22 +274,560 @@ class ModuleFusedSDPA(torch.nn.Module):
                 recompute_mode,
                 valid_sequence_lengths,
                 padding_side,
-                False,
-                False,
-                window_size)
-        else:
-            return self._hpu_kernel_fsdpa.apply(
+            )
+
+
+class QKVSliceSDPA(torch.autograd.Function):
+    '''
+    APC: split Prefix and Causal SDPA and do QKV slice on both parts.
+    '''
+
+    @staticmethod
+    def forward(ctx,
+                query,
+                key_prefix,
+                value_prefix,
+                key,
+                value,
+                causal_attn_mask,
+                q_chunk_size,
+                kv_chunk_size,
+                causal_chunk_size,
+                scale,
+                softmax_mode,
+                valid_seq_len=None,
+                padding_mode="right"):
+
+        prefix_out, prefix_m, prefix_linv = QKVSlicePrefixSDPA.apply(
+            query,
+            key_prefix,
+            value_prefix,
+            q_chunk_size,
+            kv_chunk_size,
+            scale,
+            softmax_mode,
+            valid_seq_len,
+            padding_mode,
+            False,  #output_original_dtype
+        )
+
+        text_out, text_m, text_linv = QKVSliceCausalSDPA.apply(
             query,
             key,
             value,
-            attn_mask,
-            dropout_p,
-            is_causal,
+            causal_chunk_size,
+            causal_attn_mask,
             scale,
             softmax_mode,
-            recompute_mode,
-            valid_sequence_lengths,
-            padding_side,
+            valid_seq_len,
+            padding_mode,
+            False,  #output_original_dtype
         )
 
+        new_m = torch.maximum(prefix_m, text_m)
+        prefix_linv_rescaled = (1.0 / prefix_linv) * torch.exp(prefix_m -
+                                                               new_m)
+        text_linv_rescaled = (1.0 / text_linv) * torch.exp(text_m - new_m)
+        new_linv = 1.0 / (prefix_linv_rescaled + text_linv_rescaled)
+        output = (prefix_linv_rescaled * new_linv) * prefix_out + (
+            text_linv_rescaled * new_linv) * text_out
+        return output.to(query.dtype)
 
+
+class SplitPrefixCausalSDPA(torch.autograd.Function):
+    '''
+    APC: split Prefix and Causal SDPA parts.
+    '''
+
+    @staticmethod
+    def forward(ctx,
+                query,
+                key_prefix,
+                value_prefix,
+                key,
+                value,
+                causal_attn_mask,
+                q_chunk_size,
+                kv_chunk_size,
+                causal_chunk_size,
+                scale,
+                softmax_mode,
+                valid_seq_len=None,
+                padding_mode="right"):
+
+        prefix_out, prefix_m, prefix_linv = PrefixSDPA.apply(
+            query,
+            key_prefix,
+            value_prefix,
+            scale,
+            softmax_mode,
+            valid_seq_len,
+            padding_mode,
+            False,  #output_original_dtype
+        )
+
+        text_out, text_m, text_linv = CausalSDPA.apply(
+            query,
+            key,
+            value,
+            causal_attn_mask,
+            scale,
+            softmax_mode,
+            valid_seq_len,
+            padding_mode,
+            False,  #output_original_dtype
+        )
+
+        new_m = torch.maximum(prefix_m, text_m)
+        prefix_linv_rescaled = (1.0 / prefix_linv) * torch.exp(prefix_m -
+                                                               new_m)
+        text_linv_rescaled = (1.0 / text_linv) * torch.exp(text_m - new_m)
+        new_linv = 1.0 / (prefix_linv_rescaled + text_linv_rescaled)
+        output = (prefix_linv_rescaled * new_linv) * prefix_out + (
+            text_linv_rescaled * new_linv) * text_out
+
+        return output.to(query.dtype)
+
+
+class CausalSliceSDPA(torch.autograd.Function):
+    '''
+    APC: split Prefix and Causal SDPA and do QKV slice on Causal part only.
+    '''
+
+    @staticmethod
+    def forward(ctx,
+                query,
+                key_prefix,
+                value_prefix,
+                key,
+                value,
+                causal_attn_mask,
+                q_chunk_size,
+                kv_chunk_size,
+                causal_chunk_size,
+                scale,
+                softmax_mode,
+                valid_seq_len=None,
+                padding_mode="right"):
+
+        prefix_out, prefix_m, prefix_linv = PrefixSDPA.apply(
+            query,
+            key_prefix,
+            value_prefix,
+            scale,
+            softmax_mode,
+            valid_seq_len,
+            padding_mode,
+            False,  #output_original_dtype
+        )
+
+        text_out, text_m, text_linv = QKVSliceCausalSDPA.apply(
+            query,
+            key,
+            value,
+            causal_chunk_size,
+            causal_attn_mask,
+            scale,
+            softmax_mode,
+            valid_seq_len,
+            padding_mode,
+            False,  #output_original_dtype
+        )
+
+        new_m = torch.maximum(prefix_m, text_m)
+        prefix_linv_rescaled = (1.0 / prefix_linv) * torch.exp(prefix_m -
+                                                               new_m)
+        text_linv_rescaled = (1.0 / text_linv) * torch.exp(text_m - new_m)
+        new_linv = 1.0 / (prefix_linv_rescaled + text_linv_rescaled)
+        output = (prefix_linv_rescaled * new_linv) * prefix_out + (
+            text_linv_rescaled * new_linv) * text_out
+
+        return output.to(query.dtype)
+
+
+class PrefixSDPA(torch.autograd.Function):
+    '''
+    APC: Prefix part.
+    '''
+
+    @staticmethod
+    def forward(ctx,
+                query,
+                key_prefix,
+                value_prefix,
+                scale,
+                softmax_mode,
+                valid_seq_len=None,
+                padding_mode="right",
+                output_original_dtype=False):
+
+        gqa = is_gqa(query, key_prefix)
+        if gqa:
+            query, key_prefix, value_prefix, _ = gqa_input_reshape_fwd(
+                query, key_prefix, value_prefix, None)
+
+        result = torch.ops.hpu.sdpa_recomp_fwd(
+            query,
+            key_prefix,
+            value_prefix,
+            None,  #attn_mask
+            0.0,  #dropout
+            scale,
+            False,  #is_causal
+            True,  #requires_backward
+            softmax_mode,
+            valid_seq_len,
+            padding_mode)
+
+        out, m, linv = (gqa_output_reshape(x) if gqa else x
+                        for x in result[:3])
+
+        if output_original_dtype:
+            out = out.to(query.dtype)
+        else:
+            out = out.to(torch.float32)
+        m = m.to(torch.float32)
+        if softmax_mode == "fast":
+            linv = linv.to(torch.float32) * 128.0
+        else:
+            linv = linv.to(torch.float32)
+
+        return out, m, linv
+
+
+class QKVSlicePrefixSDPA(torch.autograd.Function):
+    '''
+    APC: Prefix part with QKV slice.
+    '''
+
+    @staticmethod
+    def forward(ctx,
+                query,
+                key_prefix,
+                value_prefix,
+                q_chunk_size,
+                kv_chunk_size,
+                scale,
+                softmax_mode,
+                valid_seq_len=None,
+                padding_mode="right",
+                output_original_dtype=False):
+
+        gqa = is_gqa(query, key_prefix)
+        if gqa:
+            query, key_prefix, value_prefix, _ = gqa_input_reshape_fwd(
+                query, key_prefix, value_prefix, None)
+
+        query_len = query.size(-2)
+        num_query_chunk = (query_len + q_chunk_size - 1) // q_chunk_size
+
+        key_len = key_prefix.size(-2)
+        num_kv_chunk = (key_len + kv_chunk_size - 1) // kv_chunk_size
+
+        final_hidden_list = []
+        final_m_list = []
+        final_linv_list = []
+
+        for query_idx in range(num_query_chunk):
+
+            query_start = query_idx * q_chunk_size
+            query_end = min((query_idx + 1) * q_chunk_size, query_len)
+
+            query_slice = query[..., query_start:query_end, :]
+
+            out = None
+            m = None
+            linv = None
+
+            for kv_idx in range(num_kv_chunk):
+
+                kv_start = kv_idx * kv_chunk_size
+                kv_end = min((kv_idx + 1) * kv_chunk_size, key_len)
+
+                key_slice = key_prefix[..., kv_start:kv_end, :]
+                value_slice = value_prefix[..., kv_start:kv_end, :]
+
+                block_result = torch.ops.hpu.sdpa_recomp_fwd(
+                    query_slice,
+                    key_slice,
+                    value_slice,
+                    None,  #attn_mask
+                    0.0,  #dropout
+                    scale,
+                    False,  #is_causal
+                    True,  #requires_backward
+                    softmax_mode,
+                    valid_seq_len,
+                    padding_mode)
+
+                block_out, block_m, block_linv = (gqa_output_reshape(x)
+                                                  if gqa else x
+                                                  for x in block_result[:3])
+                block_out = block_out.to(torch.float32)
+                block_m = block_m.to(torch.float32)
+                if softmax_mode == "fast":
+                    block_linv = block_linv.to(torch.float32) * 128.0
+                else:
+                    block_linv = block_linv.to(torch.float32)
+
+                if kv_idx == 0:
+                    out = block_out
+                    m = block_m
+                    linv = block_linv
+                else:
+                    new_m = torch.maximum(m, block_m)
+                    l_rescaled = (1.0 / linv) * torch.exp(m - new_m)
+                    block_l_rescaled = (1.0 / block_linv) * torch.exp(block_m -
+                                                                      new_m)
+                    new_linv = 1.0 / (l_rescaled + block_l_rescaled)
+                    out = (l_rescaled * new_linv) * out + (
+                        block_l_rescaled * new_linv) * block_out
+                    linv = new_linv
+                    m = new_m
+
+            if output_original_dtype:
+                final_hidden_list.append(out.to(query.dtype))
+            else:
+                final_hidden_list.append(out)
+            final_linv_list.append(linv)
+            final_m_list.append(m)
+
+        output = torch.cat(final_hidden_list, dim=-2)
+        final_linv = torch.cat(final_linv_list, dim=-2)
+        final_m = torch.cat(final_m_list, dim=-2)
+
+        return output, final_m, final_linv
+
+
+class CausalSDPA(torch.autograd.Function):
+    '''
+    APC: Causal part.
+    '''
+
+    @staticmethod
+    def forward(ctx,
+                query,
+                key,
+                value,
+                attn_mask,
+                scale,
+                softmax_mode,
+                valid_seq_len=None,
+                padding_mode="right",
+                output_original_dtype=False):
+
+        query_len = query.size(-2)
+        kv_len = key.size(-2)
+
+        assert query_len == kv_len, f"For causal SDPA, {query_len=} should be equal to {kv_len=}"
+
+        gqa = is_gqa(query, key)
+        if gqa:
+            query, key, value, attn_mask = gqa_input_reshape_fwd(
+                query, key, value, attn_mask)
+
+        # Kernel limitation, need to use not causal and pass mask to get correct m and linv
+        if query_len % 1024 != 0:
+            is_causal = False
+            if attn_mask is None:
+                bs = query.size(0)
+                mask_shape = (bs, 1, 1, query_len,
+                              query_len) if gqa else (bs, 1, query_len,
+                                                      query_len)
+                mask = (1 - torch.tril(
+                    torch.ones(
+                        mask_shape, dtype=query.dtype,
+                        device=query.device))) * torch.finfo(query.dtype).min
+            else:
+                mask = attn_mask
+        else:
+            is_causal = True
+            mask = None
+
+        result = torch.ops.hpu.sdpa_recomp_fwd(
+            query,
+            key,
+            value,
+            mask,  #attn_mask
+            0.0,  #dropout
+            scale,
+            is_causal,  #is_causal
+            True,  #requires_backward
+            softmax_mode,
+            valid_seq_len,
+            padding_mode)
+
+        out, m, linv = (gqa_output_reshape(x) if gqa else x
+                        for x in result[:3])
+
+        if output_original_dtype:
+            out = out.to(query.dtype)
+        else:
+            out = out.to(torch.float32)
+        m = m.to(torch.float32)
+        if softmax_mode == "fast":
+            linv = linv.to(torch.float32) * 128.0
+        else:
+            linv = linv.to(torch.float32)
+
+        return out, m, linv
+
+
+class QKVSliceCausalSDPA(torch.autograd.Function):
+    '''
+    APC: Causal part with QKV slice.
+    '''
+
+    @staticmethod
+    def forward(ctx,
+                query,
+                key,
+                value,
+                chunk_size,
+                attn_mask,
+                scale,
+                softmax_mode,
+                valid_seq_len=None,
+                padding_mode="right",
+                output_original_dtype=False):
+
+        query_len = query.size(-2)
+        kv_len = key.size(-2)
+
+        assert query_len == kv_len, f"For QKV slice causal SDPA, {query_len=} should be equal to {kv_len=}"
+
+        gqa = is_gqa(query, key)
+        if gqa:
+            query, key, value, attn_mask = gqa_input_reshape_fwd(
+                query, key, value, attn_mask)
+
+        num_chunk = (query_len + chunk_size - 1) // chunk_size
+
+        final_hidden_list = []
+        final_m_list = []
+        final_linv_list = []
+
+        for query_idx in range(num_chunk):
+
+            query_start = query_idx * chunk_size
+            query_end = min((query_idx + 1) * chunk_size, query_len)
+
+            # step 1: compute the diagonal casual blocks
+            query_slice = query[..., query_start:query_end, :]
+            key_slice = key[..., query_start:query_end, :]
+            value_slice = value[..., query_start:query_end, :]
+
+            bs = query_slice.size(0)
+            q_slice_len = query_slice.size(-2)
+
+            # Kernel limitation, need to use not causal and pass mask to get correct m and linv
+            if q_slice_len % 1024 != 0 or q_slice_len < chunk_size:
+                if attn_mask is not None:
+                    mask = attn_mask[..., query_start:query_end,
+                                     query_start:query_end]
+                else:
+                    mask_shape = (bs, 1, 1, q_slice_len,
+                                  q_slice_len) if gqa else (bs, 1, q_slice_len,
+                                                            q_slice_len)
+                    mask = (1 - torch.tril(
+                        torch.ones(mask_shape,
+                                   dtype=query.dtype,
+                                   device=query.device))) * torch.finfo(
+                                       query.dtype).min
+
+                result = torch.ops.hpu.sdpa_recomp_fwd(
+                    query_slice,
+                    key_slice,
+                    value_slice,
+                    mask,
+                    0.0,  #dropout
+                    scale,
+                    False,  #is_causal
+                    True,  #requires_backward
+                    softmax_mode,
+                    None,  #valid_seq_len
+                    padding_mode)
+            else:
+                # causal
+                result = torch.ops.hpu.sdpa_recomp_fwd(
+                    query_slice,
+                    key_slice,
+                    value_slice,
+                    None,  #mask
+                    0.0,  #dropout
+                    scale,
+                    True,  #is_causal
+                    True,  #requires_backward
+                    softmax_mode,
+                    None,  #valid_seq_len
+                    padding_mode)
+
+            out, m, linv = (gqa_output_reshape(x) if gqa else x
+                            for x in result[:3])
+
+            out = out.to(torch.float32)
+            m = m.to(torch.float32)
+            if softmax_mode == "fast":
+                linv = linv.to(torch.float32) * 128.0
+            else:
+                linv = linv.to(torch.float32)
+
+            if num_chunk == 1:
+                if output_original_dtype:
+                    out = out.to(query.dtype)
+                return out, m, linv
+
+            # step 2: compute the full attn blocks
+            for kv_idx in range(0, query_idx):
+                kv_start = kv_idx * chunk_size
+                kv_end = (kv_idx + 1) * chunk_size
+
+                key_slice = key[..., kv_start:kv_end, :]
+                value_slice = value[..., kv_start:kv_end, :]
+
+                block_result = torch.ops.hpu.sdpa_recomp_fwd(
+                    query_slice,
+                    key_slice,
+                    value_slice,
+                    None,  #mask
+                    0.0,  #dropout
+                    scale,
+                    False,  #is_causal
+                    True,  #requires_backward
+                    softmax_mode,
+                    None,  #valid_seq_len
+                    padding_mode)
+
+                block_out, block_m, block_linv = (gqa_output_reshape(x)
+                                                  if gqa else x
+                                                  for x in block_result[:3])
+
+                block_out = block_out.to(torch.float32)
+                block_m = block_m.to(torch.float32)
+                if softmax_mode == "fast":
+                    block_linv = block_linv.to(torch.float32) * 128.0
+                else:
+                    block_linv = block_linv.to(torch.float32)
+
+                new_m = torch.maximum(m, block_m)
+                l_rescaled = (1.0 / linv) * torch.exp(m - new_m)
+                block_l_rescaled = (1.0 / block_linv) * torch.exp(block_m -
+                                                                  new_m)
+                new_linv = 1.0 / (l_rescaled + block_l_rescaled)
+                out = (l_rescaled * new_linv) * out + (block_l_rescaled *
+                                                       new_linv) * block_out
+                linv = new_linv
+                m = new_m
+
+            if output_original_dtype:
+                final_hidden_list.append(out.to(query.dtype))
+            else:
+                final_hidden_list.append(out)
+            final_linv_list.append(linv)
+            final_m_list.append(m)
+
+        output = torch.cat(final_hidden_list, dim=-2)
+        final_linv = torch.cat(final_linv_list, dim=-2)
+        final_m = torch.cat(final_m_list, dim=-2)
+
+        return output, final_m, final_linv
